Fix Linux x86_64 nightly build. Fix and set up to date Android builds.

+++ include/eepp/system/iostream.hpp
@@ -4,7 +4,7 @@
 #include <eepp/config.hpp>
 
 namespace EE {
-using ios_size = long;
+using ios_size = long long;
 }
 
 namespace EE { namespace System {
